+++ lib/src/logger/logger.dart
@@ -1,0 +1,4 @@
+/// Copyright (c) 2025 Wojciech Plesiak
+
+export 'dlog_writer.dart';
+export 'dlog.dart';